requirements.txt
flask-restful==0.3.10
flask-cors==5.0.1
azure-cosmos==4.9.0
werkzeug==3.1.3
azure-identity==1.21.0
authlib==1.5.1
flask-httpauth==4.8.0
openai==0.28.0
pandas==2.2.3
openpyxl==3.1.5
PyMuPDF==1.25.3
pillow==11.1.0
python-docx==1.1.2
pdfplumber==0.11.5
chardet==5.2.0
azure-storage-blob==12.25.0
opencensus-ext-azure==1.1.14
xlrd==2.0.1
